git-paid/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tinyfish                # Official TinyFish Python SDK
python-dotenv==1.0.1
pydantic==2.10.4